feat(38): 리뷰 목록 조회 기능 구현 (#50)
* ♻️refactor: 평점 타입 변경
- rating 타입 int에서 double로 변경

* ♻️ refactor: rating 타입 변경

* ✨feat: 리뷰 목록 조회 기능 구현
- 리뷰 목록을 다양한 조건으로 조회하는 기능 추가
- 정렬 기준 (최신순, 오래된순, 평점 높은순, 평점 낮은순)
- 제목 검색, 여행지별, 회원별 리뷰 조회

+++ backend/src/main/java/com/tripfriend/domain/review/entity/Review.java
@@ -26,7 +26,7 @@ public class Review {
     private String content;
 
     @Column(nullable = false)
-    private int rating;
+    private double rating;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
@@ -41,7 +41,7 @@ public class Review {
     @Column(nullable = false)
     private LocalDateTime updatedAt;
 
-    public Review(String title, String content, int rating, Member member, Long placeId) {
+    public Review(String title, String content, double rating, Member member, Long placeId) {
         this.title = title;
         this.content = content;
         this.rating = rating;
@@ -51,7 +51,7 @@ public class Review {
         this.updatedAt = this.createdAt;
     }
 
-    public void update(String title, String content, int rating) {
+    public void update(String title, String content, double rating) {
         this.title = title;
         this.content = content;
         this.rating = rating;
